Factor out the map and hazard grouping.
